pyproject.toml
[tool.poetry]
name = "brioa_port"
version = "0.1.0"
description = ""
authors = ["yurihs"]

[tool.poetry.dependencies]
python = "^3.7"
python-dateutil = "^2.7"
schedule = "^0.5.0"
docopt = "^0.6.2"

[tool.poetry.dev-dependencies]
pytest = "^3.0"
mypy = "^0.650.0"
pytest-mock = "^1.10"

[tool.poetry.scripts]
brioa_webcam_downloader = "brioa_port.scripts.brioa_webcam_downloader:main"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
